Move js/wasm/svelte/host.ts -> js/wasm/network/host.ts
