+++ js/app/src/lite/websocket.ts
@@ -1,5 +1,5 @@
 import type { WorkerProxy } from "@gradio/wasm";
-import { is_self_host } from "@gradio/wasm/svelte";
+import { is_self_host } from "@gradio/wasm/network";
 
 /**
  * A WebSocket factory that proxies requests to the worker,
