+++ js/wasm/svelte/index.ts
@@ -1,3 +1,2 @@
 export * from "./context";
 export * from "./file-url";
-export * from "./host";
